requirements.txt
altair==4.2.2
async-generator==1.10
attrs==22.2.0
beautifulsoup4==4.11.1
blinker==1.5
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
decorator==5.1.1
docopt==0.6.2
entrypoints==0.4
exceptiongroup==1.1.0
gitdb==4.0.10
GitPython==3.1.31
greenlet==2.0.2
h11==0.14.0
idna==3.4
importlib-metadata==6.0.0
Jinja2==3.1.2
jsonschema==4.17.3
markdown-it-py==2.2.0
MarkupSafe==2.1.2
mdurl==0.1.2
mysql-connector-python==8.0.32
numpy==1.24.2
outcome==1.2.0
packaging==23.0
pandas==1.5.3
Pillow==9.4.0
pipreqs==0.4.11
plotly==5.14.1
protobuf==3.20.3
pyarrow==11.0.0
pydeck==0.8.0
Pygments==2.14.0
Pympler==1.0.1
PyMySQL==1.0.2
pyrsistent==0.19.3
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.21.1
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
requests==2.28.2
rich==13.3.2
schedule==1.1.0
selenium==4.8.0
semver==2.13.0
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
scikit-learn==1.2.2
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
SQLAlchemy==2.0.4
streamlit==1.20.0
tenacity==8.2.2
toml==0.10.2
toolz==0.12.0
tornado==6.2
tqdm==4.64.1
trio==0.22.0
trio-websocket==0.9.2
typing_extensions==4.5.0
tzdata==2022.7
tzlocal==4.2
urllib3==1.26.14
validators==0.20.0
webdriver-manager==3.8.5
wsproto==1.2.0
yarg==0.1.9
zipp==3.15.0
